Shuffle getPartnerToken
This function signature needs to be revised but that will be for later.
