+++ be/src/testing/data/webhooks.ts
@@ -1,6 +1,6 @@
 import crypto from 'crypto';
 
-import { WebhookBody } from '../partnerWebhook/types';
+import { WebhookBody } from '../../partnerWebhook/types';
 
 export const SIGNED_WEBHOOK_SUBSCRIPTION_ID = 'signed';
