+++ be/src/testing/mock.ts
@@ -1,8 +1,5 @@
-import { GetPartnerTokenRequest } from '../getPartnerToken';
+import { RetrieveRequest } from '../getPartnerToken';
 
-export const getPartnerToken = jest.fn<
-  Promise<string>,
-  [GetPartnerTokenRequest]
->();
+export const getPartnerToken = jest.fn<Promise<string>, [RetrieveRequest]>();
 
 export const getSigningSecret = jest.fn<Promise<string | null>, [string]>();
